requirements.txt
# python=3.7.16
dgl-cu102==0.4.3post2
gpustat==1.1
numba==0.56.4
numpy==1.21.6
pandas==1.3.5
scipy==1.10.1
scikit-learn==1.0.1
torch==1.9.1+cu102
tqdm==4.62.3